+++ ydb/opentelemetry/plugin.py
@@ -2,12 +2,17 @@
 
 from opentelemetry import context as otel_context
 from opentelemetry import trace
+from opentelemetry.metrics import Observation
+from opentelemetry import metrics
 from opentelemetry.propagate import inject
 from opentelemetry.trace import StatusCode
 
 from ydb import issues
 from ydb.issues import StatusCode as YdbStatusCode
-from ydb.opentelemetry.tracing import _registry
+from ydb.opentelemetry import metrics
+
+from ydb.opentelemetry.metrics import _metrics_registry, create_metrics_operation
+from ydb.opentelemetry.tracing import _registry as _tracing_registry
 
 # YDB client transport StatusCode values (401xxx band) -> OTel error.type transport_error.
 _TRANSPORT_STATUSES = frozenset(
@@ -21,7 +26,8 @@ _TRANSPORT_STATUSES = frozenset(
 )
 
 _tracer = None
-_enabled = False
+_tracing_enabled = False
+_meter = None
 
 _KIND_MAP = {
     "client": trace.SpanKind.CLIENT,
@@ -113,22 +119,55 @@ def _create_span(name, attributes=None, kind=None):
 
 
 def _enable_tracing(tracer=None):
-    global _enabled, _tracer
+    global _tracing_enabled, _tracer
 
-    if _enabled:
+    if _tracing_enabled:
         return
 
     _tracer = tracer if tracer is not None else trace.get_tracer("ydb.sdk")
-    _enabled = True
-    _registry.set_metadata_hook(_otel_metadata_hook)
-    _registry.set_create_span(_create_span)
+    _tracing_enabled = True
+    _tracing_registry.set_metadata_hook(_otel_metadata_hook)
+    _tracing_registry.set_create_span(_create_span)
 
 
 def _disable_tracing():
     """Clear hooks and tracer; after this, :func:`~ydb.opentelemetry.enable_tracing` may be called again."""
-    global _enabled, _tracer
+    global _tracing_enabled, _tracer
 
-    _registry.set_create_span(None)
-    _registry.set_metadata_hook(None)
-    _enabled = False
+    _tracing_registry.set_create_span(None)
+    _tracing_registry.set_metadata_hook(None)
+    _tracing_enabled = False
     _tracer = None
+
+def _create_query_session_count_callback():
+    """Create callback for observable query session count metric."""
+
+    def observe_query_session_count(_):
+        values = _metrics_registry.get_query_session_count_values()
+        return [Observation(value, attributes=dict(attrs)) for attrs, value in values.items()]
+
+    return observe_query_session_count
+
+
+def _enable_metrics(meter_provider):
+    global _meter
+
+    if _meter is not None:
+        return
+
+    if meter_provider is None:
+        _meter = metrics.get_meter("ydb.sdk")
+    elif hasattr(meter_provider, "get_meter"):
+        _meter = meter_provider.get_meter("ydb.sdk")
+    else:
+        raise TypeError("meter_provider must be an OpenTelemetry MeterProvider")
+
+    _metrics_registry.set_meter(_meter, _create_query_session_count_callback())
+
+
+def _disable_metrics():
+    global _meter
+
+    _metrics_registry.clear()
+    if _meter is not None:
+        _meter = None
